+++ src/store/index.js
@@ -6,6 +6,7 @@ import accounts from './accounts';
 import accountDetails from './accountDetails';
 import currentAccountOperation from './currentAccountOperation';
 import periodicOperation from './periodicOperation';
+import publicService from './publicService';
 
 Vue.use(Vuex);
 
@@ -70,6 +71,6 @@ export default new Vuex.Store({
     },
 
 	modules: {
-		auth, snackbar, accounts, accountDetails, currentAccountOperation, periodicOperation
+		auth, snackbar, accounts, accountDetails, currentAccountOperation, periodicOperation, publicService
 	},
 });